requirements.txt
asgiref==3.8.1
attrs==24.2.0
boto3==1.35.44
botocore==1.35.44
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chromedriver-autoinstaller==0.6.4
cryptography==43.0.3
dj-database-url==2.2.0
Django==5.1.2
django-appconf==1.0.6
django-cors-headers==4.5.0
django-cryptography==1.1
django-encrypted-model-fields==0.6.5
django-environ==0.11.2
django-extensions==3.2.3
django-ses==4.2.0
djangorestframework==3.15.2
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
MarkupSafe==3.0.2
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pillow==11.0.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg==3.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.4
whitenoise==6.7.0
wsproto==1.2.0
